requirements.txt
python-dotenv
pydantic

# Scraping
requests
bs4
lxml

# Scheduler
apscheduler
sqlalchemy

# DAO
supabase

# Arxiv Collector
feedparser

# Zotero Collector
pyzotero